refactor: file organisation — splits and renames
- Delete copyTree wrapper, callers use assembler.CopyDir directly
- Rename helpers.go → install_helpers.go (remaining functions are install-related)
- Split inspect.go (847 LOC) into inspect.go (routing, args, interactive)
  and inspect_analyze.go (LLM prompts, analysis, parsing)
- Split exporter_test.go by vendor: claude, codex, merged
- Split resolver_test.go: extract cache-related tests to resolver_cache_test.go
- Split marketplace_test.go: extract build and index tests
- Add copydir_test.go for assembler.CopyDir (moved from helpers_test.go)
