+++ cmd/ynh/main.go
@@ -243,7 +243,7 @@ func cmdInstall(args []string) error {
 		return fmt.Errorf("creating install directory: %w", err)
 	}
 
-	if err := copyTree(srcDir, installDir); err != nil {
+	if err := assembler.CopyDir(srcDir, installDir); err != nil {
 		return fmt.Errorf("copying harness to install directory: %w", err)
 	}
