+++ internal/resolver/resolver_test.go
@@ -361,105 +361,6 @@ func TestResolve_WithPath_NoPickIncludesAll(t *testing.T) {
 	}
 }
 
-func TestEnsureRepo_CacheUpdatesWorkingTree(t *testing.T) {
-	// Create a local git repo
-	srcDir := t.TempDir()
-	runGit(t, srcDir, "init")
-	runGit(t, srcDir, "config", "user.email", "test@test.com")
-	runGit(t, srcDir, "config", "user.name", "Test")
-
-	if err := os.WriteFile(filepath.Join(srcDir, "file.txt"), []byte("v1"), 0o644); err != nil {
-		t.Fatal(err)
-	}
-	runGit(t, srcDir, "add", ".")
-	runGit(t, srcDir, "commit", "-m", "v1")
-
-	t.Setenv("YNH_HOME", "")
-	t.Setenv("HOME", t.TempDir())
-
-	// First clone
-	result, err := EnsureRepo(srcDir, "")
-	if err != nil {
-		t.Fatalf("first ensureRepo failed: %v", err)
-	}
-	data, err := os.ReadFile(filepath.Join(result.Path, "file.txt"))
-	if err != nil {
-		t.Fatal(err)
-	}
-	if string(data) != "v1" {
-		t.Fatalf("expected v1, got %q", string(data))
-	}
-
-	// Update source repo
-	if err := os.WriteFile(filepath.Join(srcDir, "file.txt"), []byte("v2"), 0o644); err != nil {
-		t.Fatal(err)
-	}
-	runGit(t, srcDir, "add", ".")
-	runGit(t, srcDir, "commit", "-m", "v2")
-
-	// Second call should update working tree
-	result2, err := EnsureRepo(srcDir, "")
-	if err != nil {
-		t.Fatalf("second ensureRepo failed: %v", err)
-	}
-	if !result2.Changed {
-		t.Error("expected Changed=true after upstream commit")
-	}
-
-	data, err = os.ReadFile(filepath.Join(result2.Path, "file.txt"))
-	if err != nil {
-		t.Fatal(err)
-	}
-	if string(data) != "v2" {
-		t.Errorf("working tree not updated: got %q, want %q", string(data), "v2")
-	}
-}
-
-func TestEnsureRepo_UpdateErrorsNotSwallowed(t *testing.T) {
-	// Create a local git repo and clone it
-	srcDir := t.TempDir()
-	runGit(t, srcDir, "init")
-	runGit(t, srcDir, "config", "user.email", "test@test.com")
-	runGit(t, srcDir, "config", "user.name", "Test")
-	if err := os.WriteFile(filepath.Join(srcDir, "file.txt"), []byte("v1"), 0o644); err != nil {
-		t.Fatal(err)
-	}
-	runGit(t, srcDir, "add", ".")
-	runGit(t, srcDir, "commit", "-m", "v1")
-
-	t.Setenv("YNH_HOME", "")
-	t.Setenv("HOME", t.TempDir())
-
-	// First call: clone succeeds
-	result, err := EnsureRepo(srcDir, "")
-	if err != nil {
-		t.Fatalf("initial clone failed: %v", err)
-	}
-	_ = result
-
-	// Corrupt the remote by removing the source repo's .git
-	if err := os.RemoveAll(filepath.Join(srcDir, ".git")); err != nil {
-		t.Fatal(err)
-	}
-
-	// Second call: update should fail (fetch from a non-git directory)
-	_, err = EnsureRepo(srcDir, "")
-	if err == nil {
-		t.Fatal("expected error when fetching from corrupted remote, got nil")
-	}
-
-	// Verify the error message contains useful context
-	errStr := err.Error()
-	if !strings.Contains(errStr, "git fetch") {
-		t.Errorf("error should mention git fetch, got: %s", errStr)
-	}
-
-	// Verify the cached repo still exists (we don't blow it away on update failure)
-	if _, statErr := os.Stat(result.Path); os.IsNotExist(statErr) {
-		t.Error("cached repo should still exist after failed update")
-	}
-}
-
 func TestResolve_BlockedByAllowList(t *testing.T) {
 	cfg := &config.Config{
 		AllowedRemoteSources: []string{
@@ -561,253 +462,6 @@ func TestResolve_NilConfigAllowsAll(t *testing.T) {
 	}
 	if len(results) != 0 {
 		t.Errorf("expected 0 results for empty includes, got %d", len(results))
-	}
-}
-
-func TestCacheOnlyRepo_CacheHit(t *testing.T) {
-	// Create a local git repo and clone it via EnsureRepo first
-	srcDir := t.TempDir()
-	runGit(t, srcDir, "init")
-	runGit(t, srcDir, "config", "user.email", "test@test.com")
-	runGit(t, srcDir, "config", "user.name", "Test")
-	if err := os.WriteFile(filepath.Join(srcDir, "test.txt"), []byte("cached"), 0o644); err != nil {
-		t.Fatal(err)
-	}
-	runGit(t, srcDir, "add", ".")
-	runGit(t, srcDir, "commit", "-m", "init")
-
-	t.Setenv("YNH_HOME", "")
-	t.Setenv("HOME", t.TempDir())
-
-	// Populate cache via EnsureRepo
-	ensureResult, err := EnsureRepo(srcDir, "")
-	if err != nil {
-		t.Fatalf("EnsureRepo failed: %v", err)
-	}
-
-	// Now remove the source repo so any network fetch would fail
-	if err := os.RemoveAll(filepath.Join(srcDir, ".git")); err != nil {
-		t.Fatal(err)
-	}
-
-	// CacheOnlyRepo should return cached result without hitting the network
-	result, err := CacheOnlyRepo(srcDir, "")
-	if err != nil {
-		t.Fatalf("CacheOnlyRepo should succeed on cache hit: %v", err)
-	}
-	if result.Path != ensureResult.Path {
-		t.Errorf("expected same path %q, got %q", ensureResult.Path, result.Path)
-	}
-	if result.Cloned {
-		t.Error("expected Cloned=false for cache hit")
-	}
-
-	// Verify content is still accessible
-	data, err := os.ReadFile(filepath.Join(result.Path, "test.txt"))
-	if err != nil {
-		t.Fatalf("cached file not readable: %v", err)
-	}
-	if string(data) != "cached" {
-		t.Errorf("cached content = %q, want %q", string(data), "cached")
-	}
-}
-
-func TestCacheOnlyRepo_CacheMiss(t *testing.T) {
-	// Create a local git repo — cache is empty, so CacheOnlyRepo should clone
-	srcDir := t.TempDir()
-	runGit(t, srcDir, "init")
-	runGit(t, srcDir, "config", "user.email", "test@test.com")
-	runGit(t, srcDir, "config", "user.name", "Test")
-	if err := os.WriteFile(filepath.Join(srcDir, "test.txt"), []byte("fresh"), 0o644); err != nil {
-		t.Fatal(err)
-	}
-	runGit(t, srcDir, "add", ".")
-	runGit(t, srcDir, "commit", "-m", "init")
-
-	t.Setenv("YNH_HOME", "")
-	t.Setenv("HOME", t.TempDir())
-
-	// CacheOnlyRepo with empty cache should fall back to EnsureRepo
-	result, err := CacheOnlyRepo(srcDir, "")
-	if err != nil {
-		t.Fatalf("CacheOnlyRepo should fall back to clone: %v", err)
-	}
-	if !result.Cloned {
-		t.Error("expected Cloned=true for cache miss fallback")
-	}
-
-	data, err := os.ReadFile(filepath.Join(result.Path, "test.txt"))
-	if err != nil {
-		t.Fatalf("cloned file not readable: %v", err)
-	}
-	if string(data) != "fresh" {
-		t.Errorf("cloned content = %q, want %q", string(data), "fresh")
-	}
-}
-
-func TestResolveFromCache_UsesCache(t *testing.T) {
-	// Create a local git repo with skills
-	srcDir := t.TempDir()
-	runGit(t, srcDir, "init")
-	runGit(t, srcDir, "config", "user.email", "test@test.com")
-	runGit(t, srcDir, "config", "user.name", "Test")
-	if err := os.MkdirAll(filepath.Join(srcDir, "skills", "hello"), 0o755); err != nil {
-		t.Fatal(err)
-	}
-	if err := os.WriteFile(filepath.Join(srcDir, "skills", "hello", "SKILL.md"), []byte("hello"), 0o644); err != nil {
-		t.Fatal(err)
-	}
-	runGit(t, srcDir, "add", ".")
-	runGit(t, srcDir, "commit", "-m", "init")
-
-	t.Setenv("YNH_HOME", "")
-	t.Setenv("HOME", t.TempDir())
-
-	// Pre-populate cache
-	if _, err := EnsureRepo(srcDir, ""); err != nil {
-		t.Fatalf("EnsureRepo failed: %v", err)
-	}
-
-	// Remove source so network access would fail
-	if err := os.RemoveAll(filepath.Join(srcDir, ".git")); err != nil {
-		t.Fatal(err)
-	}
-
-	p := &harness.Harness{
-		Name: "test-cached",
-		Includes: []harness.Include{
-			{
-				GitSource: harness.GitSource{Git: srcDir},
-				Pick:      []string{"skills/hello"},
-			},
-		},
-	}
-
-	// ResolveFromCache should succeed from cache alone
-	results, err := ResolveFromCache(p, nil)
-	if err != nil {
-		t.Fatalf("ResolveFromCache failed: %v", err)
-	}
-	if len(results) != 1 {
-		t.Fatalf("expected 1 result, got %d", len(results))
-	}
-	if results[0].Cloned {
-		t.Error("expected Cloned=false (served from cache)")
-	}
-	if !results[0].Cached {
-		t.Error("expected Cached=true")
-	}
-}
-
-func TestResolveFromCache_BlockedByAllowList(t *testing.T) {
-	cfg := &config.Config{
-		AllowedRemoteSources: []string{
-			"github.com/trusted-org/*",
-		},
-	}
-
-	p := &harness.Harness{
-		Name: "test-blocked-cache",
-		Includes: []harness.Include{
-			{
-				GitSource: harness.GitSource{Git: "github.com/untrusted-org/repo"},
-			},
-		},
-	}
-
-	_, err := ResolveFromCache(p, cfg)
-	if err == nil {
-		t.Fatal("expected error for blocked remote source")
-	}
-	if !strings.Contains(err.Error(), "not in the allowed sources list") {
-		t.Errorf("error should mention allow list, got: %v", err)
-	}
-}
-
-func TestResolveGitSourceFromCache_WithPath(t *testing.T) {
-	// Create a repo with a subdirectory
-	srcDir := t.TempDir()
-	runGit(t, srcDir, "init")
-	runGit(t, srcDir, "config", "user.email", "test@test.com")
-	runGit(t, srcDir, "config", "user.name", "Test")
-	if err := os.MkdirAll(filepath.Join(srcDir, "sub", "skills", "s1"), 0o755); err != nil {
-		t.Fatal(err)
-	}
-	if err := os.WriteFile(filepath.Join(srcDir, "sub", "skills", "s1", "SKILL.md"), []byte("s1"), 0o644); err != nil {
-		t.Fatal(err)
-	}
-	runGit(t, srcDir, "add", ".")
-	runGit(t, srcDir, "commit", "-m", "init")
-
-	t.Setenv("YNH_HOME", "")
-	t.Setenv("HOME", t.TempDir())
-
-	// Pre-populate cache
-	if _, err := EnsureRepo(srcDir, ""); err != nil {
-		t.Fatalf("EnsureRepo failed: %v", err)
-	}
-
-	// Resolve with valid path
-	gs := harness.GitSource{Git: srcDir, Path: "sub"}
-	basePath, _, err := ResolveGitSourceFromCache(gs)
-	if err != nil {
-		t.Fatalf("ResolveGitSourceFromCache failed: %v", err)
-	}
-	skillPath := filepath.Join(basePath, "skills", "s1", "SKILL.md")
-	if _, err := os.Stat(skillPath); err != nil {
-		t.Errorf("skill not found at resolved path: %v", err)
-	}
-
-	// Resolve with invalid path
-	gsBad := harness.GitSource{Git: srcDir, Path: "nonexistent"}
-	_, _, err = ResolveGitSourceFromCache(gsBad)
-	if err == nil {
-		t.Fatal("expected error for nonexistent path")
-	}
-}
-
-func TestResolveFromCache_WithPath(t *testing.T) {
-	// Create a monorepo-style repo
-	srcDir := t.TempDir()
-	runGit(t, srcDir, "init")
-	runGit(t, srcDir, "config", "user.email", "test@test.com")
-	runGit(t, srcDir, "config", "user.name", "Test")
-	if err := os.MkdirAll(filepath.Join(srcDir, "pkg", "skills", "deploy"), 0o755); err != nil {
-		t.Fatal(err)
-	}
-	if err := os.WriteFile(filepath.Join(srcDir, "pkg", "skills", "deploy", "SKILL.md"), []byte("deploy"), 0o644); err != nil {
-		t.Fatal(err)
-	}
-	runGit(t, srcDir, "add", ".")
-	runGit(t, srcDir, "commit", "-m", "init")
-
-	t.Setenv("YNH_HOME", "")
-	t.Setenv("HOME", t.TempDir())
-
-	// Pre-populate cache
-	if _, err := EnsureRepo(srcDir, ""); err != nil {
-		t.Fatalf("EnsureRepo failed: %v", err)
-	}
-
-	p := &harness.Harness{
-		Name: "test-path-cache",
-		Includes: []harness.Include{
-			{
-				GitSource: harness.GitSource{Git: srcDir, Path: "pkg"},
-				Pick:      []string{"skills/deploy"},
-			},
-		},
-	}
-
-	results, err := ResolveFromCache(p, nil)
-	if err != nil {
-		t.Fatalf("ResolveFromCache failed: %v", err)
-	}
-	if len(results) != 1 {
-		t.Fatalf("expected 1 result, got %d", len(results))
-	}
-	if results[0].Path != "pkg" {
-		t.Errorf("expected Path=%q, got %q", "pkg", results[0].Path)
 	}
 }
